build: 默认纯 Rust 构建
